requirements.txt
faiss-cpu
huggingface-hub
numpy
pandas
pymupdf
python-dotenv
sentence-transformers
streamlit